fix: rename file and close scanner
